Writes oriented camera info, and scene to POVRay
Scene orientation can be exported pressing the mousewheel.
The pov file is correcly oriented althout sometimes
vtkPOVExporter generates a non runnable .pov file.
Scrolling the mouse wheel has no effect since it is seen as a
zoom, not as a camera (or focal point) translation.
The VRML 2.0 file can be imported with Blender but the
orientation is not the optimum, but the initial one.
OOGL, OBJ files could still not been opened.
Global variables (renWin and camera) definition were
needed to to add the interactor/observer.